+++ CHANGELOG.md
@@ -9,6 +9,9 @@ From v1.0.0 and on, the project will adherence strictly to Semantic Versioning.
 
 ## [Unreleased]
 
+### Added
+- "lagging" option for shrink in `co.Delay` and `co.Residual`.
+
 ### Changed
 - `phantom_padding` renamed to `residual_shrink`.